Add "Referencable" support
Adds support for a custom hint to serializers as to whether objects are
able to be referenced or if they should be inlined. While this has no
meaning in RDF proper, it can be used in e.g. JSON-LD to indicate if the
object can have an indentifier, or if it must be inlined.